+++ java/mlflow-java/src/main/java/com/databricks/mlflow/models/JavaModel.java
@@ -1,3 +1,3 @@
-package com.databricks.mlflow;
+package com.databricks.mlflow.models;
 
 public abstract class JavaModel { public abstract String predict(String input); }
