+++ java/mlflow-java/src/main/java/com/databricks/mlflow/models/ScalaModel.java
@@ -1,4 +1,4 @@
-package com.databricks.mlflow;
+package com.databricks.mlflow.models;
 
 public abstract class ScalaModel extends JavaModel {
     @Override public abstract String predict(String input);
